Wrap ffmpeg's mp3->wav conversion in a progress
This adds a requirement for ffprob, to get the length in seconds of the
audiobook. This is required in order to calculate the size of the wav
file and give an accurate progression with time estimate

The progress bar isn't as nice looking, but I haven't figured out how to
make it nicer.